update leetcode solution under 800

+++ src/main/java/javacore/algorithm/leetcode/under0800/DesignHashMap.java
@@ -25,7 +25,7 @@ public class DesignHashMap {
 	 */
 	public int get(int key) {
 		if (map.containsKey(key))
-			return map.get(key).intValue();
+			return map.get(key);
 		return -1;
 	}
